+++ actix-service/src/transform_map_init_err.rs
@@ -1,6 +1,6 @@
 use std::marker::PhantomData;
 
-use futures::{Future, IntoFuture, Poll};
+use futures::{Future, Poll};
 
 use super::Transform;
 
@@ -57,10 +57,7 @@ where
     type Future = TransformMapInitErrFuture<T, S, F, E>;
 
     fn new_transform(&self, service: S) -> Self::Future {
-        TransformMapInitErrFuture::new(
-            self.t.new_transform(service).into_future(),
-            self.f.clone(),
-        )
+        TransformMapInitErrFuture::new(self.t.new_transform(service), self.f.clone())
     }
 }
 
@@ -69,7 +66,7 @@ where
     T: Transform<S>,
     F: Fn(T::InitError) -> E,
 {
-    fut: <T::Future as IntoFuture>::Future,
+    fut: T::Future,
     f: F,
 }
 
@@ -78,7 +75,7 @@ where
     T: Transform<S>,
     F: Fn(T::InitError) -> E,
 {
-    fn new(fut: <T::Future as IntoFuture>::Future, f: F) -> Self {
+    fn new(fut: T::Future, f: F) -> Self {
         TransformMapInitErrFuture { f, fut }
     }
 }